refactor: components 폴더 common 레이어 삭제 및 임포트 경로 적용
