+++ src/components/text/InfoText.tsx
