+++ src/domain/cart/components/cart-layout/CartLayout.tsx
@@ -1,5 +1,5 @@
 import styled from "@emotion/styled";
-import { Flex } from "../../../../components/common";
+import { Flex } from "../../../../components";
 import { useCartAPIData } from "../../hooks/useCartAPIData";
 import CartCheckList from "./cart-check-list/CartCheckList";
 import CartTitle from "./cart-check-list/CartTitle";
