+++ src/domain/cart/components/cart-layout/cart-check-list/cart-check-item/CartItemCheckbox.tsx
@@ -1,4 +1,4 @@
-import CheckBox from "../../../../../../components/common/inputs/CheckBox";
+import CheckBox from "../../../../../../components/inputs/CheckBox";
 import { useOrderCartList } from "../../../../../order/hooks/useOrderCartList";
 
 function CartItemCheckbox({ cartId }: { cartId: string }) {
