+++ src/domain/order/components/pay-contents/DeliveryInfo.tsx
@@ -1,6 +1,6 @@
 import styled from "@emotion/styled";
-import { Flex } from "../../../../components/common";
-import CheckboxLabel from "../../../../components/common/inputs/CheckboxLabel";
+import { Flex } from "../../../../components";
+import CheckboxLabel from "../../../../components/inputs/CheckboxLabel";
 import { useCouponContext } from "../../../../pages/order-confirm/context/CouponProvider";
 import { useOrderCartList } from "../../hooks/useOrderCartList";
 import { calculateOrders } from "../../utils/calculateOrders";
